change rain from float to int

+++ src/Sensorchannel.h
@@ -83,15 +83,15 @@ class Sensorchannel : public OpenKNX::Channel
     uint32_t m_winddir_alarmL_last_send_millis = 0;
 
     uint32_t m_rain_last_send_millis = 0;
-    float m_rain_last_send_value = -1000;
-    float m_rain_last_recv_value = -1000;
-    float m_rain_last_recv_value_restored = -1000;
+    int32_t m_rain_last_send_value = -1000;
+    int32_t m_rain_last_recv_value = -1000;
+    int32_t m_rain_last_recv_value_restored = -1000;
     uint32_t m_rain_last_recv_millis = 0;
     uint32_t m_rain_alarmH_last_send_millis = 0;
     uint32_t m_rain_alarmL_last_send_millis = 0;
-    float m_rain_offset_value = 0;
+    int32_t m_rain_offset_value = 0;
 
-    float m_rainflow_lastvalue = -1000;
+    int32_t m_rainflow_lastvalue = -1000;
     uint32_t m_rainflow_lastvalue_millis = 0;
 
     uint32_t m_rainbool_last_send_millis = 0;
